post-process: simplify and isolate analyses 1
This has the consequence of moving the top table describing what happens
if you implemet the optimal targets. This makes sense in that the
optimal targets should be considered an optional analysis and an
heuristic. Everything above the optional analysis is actual facts, in
that they are not really interpreted upon in the same way the optimal
targets is a full-on interpretation of "what is smart to do". The goal
is now to add even more analyses that follow a very modular approach.