post-processing: make post-process a multi-command interface
Split the post-processing into a report command and correlation command.
This is because the logic of these needs to be applied at different
stages (potentially) of the fuzz-introspector process. We need to
correlate binaries to Fuzz Introspector log files right after linking
whereas we need to do report generation after running the fuzzers and
extracting coverage. These are often done in different environments
(e.g. on OSS-Fuzz).

ref: https://github.com/ossf/fuzz-introspector/issues/44