post-process: refactor HTML report strings
Split generation of HTML report into smaller components. This makes it
make flexibly, which will be used for generating HTML content at the top
of the HTML report that is based on analysis done at the end of the
generation, e.g. showing conclusions and guidance that is based on
results from the end of the analysis but the HTML content of the
guidance will be shown at the top of the report.

Ref: https://github.com/ossf/fuzz-introspector/issues/46