Rename analyses to avoid Analysis class (#739)
* Rename analyses to avoid Analysis class

All analyses have an `Analysis` class, which is slightly annoying for
python documentation (https://fuzz-introspector.readthedocs.io/) as it
shows all of them with the same name. This changes the general names to
plugin-specific names, and then has a list in analyses/__init__.py with
all of the various plugin names. This improves documentation and I think
also makes the code better.

Signed-off-by: David Korczynski <david@adalogics.com>

* nits

Signed-off-by: David Korczynski <david@adalogics.com>

* fix ordering in optional analyses

Signed-off-by: David Korczynski <david@adalogics.com>

* nit

Signed-off-by: David Korczynski <david@adalogics.com>

Signed-off-by: David Korczynski <david@adalogics.com>